Move Builder to primitives
With the BuilderExt split complete, the Builder type can now be cleanly
moved to primitives.

Move Builder to primitives and re-export in bitcoin.
